More renaming and splitting
