+++ arangod/Replication2/Storage/RocksDB/CMakeLists.txt
@@ -2,6 +2,6 @@ target_sources(arango_replication2 PRIVATE
   AsyncLogWriteBatcher.cpp
   AsyncLogWriteContext.cpp
   LogPersistor.cpp
-  PersistedLog.cpp
+  LogStorageMethods.cpp
   StatePersistor.cpp
   )
