+++ arangod/Replication2/Storage/RocksDB/LogStorageMethods.h
@@ -23,56 +23,10 @@
 
 #pragma once
 
-#include "Replication2/ReplicatedLog/LogEntries.h"
-#include "Replication2/Storage/ILogPersistor.h"
-#include "Replication2/Storage/IStatePersistor.h"
 #include "Replication2/Storage/IStorageEngineMethods.h"
 #include "Replication2/Storage/RocksDB/AsyncLogWriteContext.h"
-#include "Metrics/CounterBuilder.h"
-#include "Metrics/GaugeBuilder.h"
-#include "Metrics/HistogramBuilder.h"
-#include "Metrics/LogScale.h"
 
-#include <function2.hpp>
-#include <Futures/Promise.h>
-
-#include <array>
-#include <variant>
 #include <memory>
-
-namespace arangodb {
-
-struct WriteBatchSizeScale {
-  using scale_t = metrics::LogScale<std::uint64_t>;
-  static scale_t scale() {
-    // values in bytes, smallest bucket is up to 1kb
-    return {scale_t::kSupplySmallestBucket, 2, 0, 1024, 16};
-  }
-};
-DECLARE_GAUGE(arangodb_replication2_rocksdb_num_persistor_worker, std::size_t,
-              "Number of threads running in the log persistor");
-DECLARE_GAUGE(arangodb_replication2_rocksdb_queue_length, std::size_t,
-              "Number of replicated log storage operations queued");
-DECLARE_HISTOGRAM(arangodb_replication2_rocksdb_write_batch_size,
-                  WriteBatchSizeScale,
-                  "Size of replicated log write batches in bytes");
-struct ApplyEntriesRttScale {
-  using scale_t = metrics::LogScale<std::uint64_t>;
-  static scale_t scale() {
-    // values in us, smallest bucket is up to 1ms, scales up to 2^16ms =~ 65s.
-    return {scale_t::kSupplySmallestBucket, 2, 0, 1'000, 16};
-  }
-};
-DECLARE_HISTOGRAM(arangodb_replication2_rocksdb_write_time,
-                  ApplyEntriesRttScale,
-                  "Replicated log batches write time[us]");
-DECLARE_HISTOGRAM(arangodb_replication2_rocksdb_sync_time, ApplyEntriesRttScale,
-                  "Replicated log batches sync time[us]");
-DECLARE_HISTOGRAM(arangodb_replication2_storage_operation_latency,
-                  ApplyEntriesRttScale,
-                  "Replicated log storage operation latency[us]");
-
-}  // namespace arangodb
 
 namespace rocksdb {
 class DB;
@@ -83,6 +37,8 @@ namespace arangodb::replication2::storage::rocksdb {
 
 struct AsyncLogWriteBatcherMetrics;
 struct IAsyncLogWriteBatcher;
+struct ILogPersistor;
+struct IStatePersistor;
 
 struct LogStorageMethods final : replication2::storage::IStorageEngineMethods {
   explicit LogStorageMethods(
